fixed decimal

+++ index.js
@@ -71,10 +71,10 @@ const runComparison = () => {
 
     leftSideStats.forEach((leftStat, index) => {
         const rightStat = rightSideStats[index];
-        const leftSideValue = parseInt(leftStat.dataset.value);
-        const rightSideValue = parseInt(rightStat.dataset.value);
+        const leftSideValue = parseFloat(leftStat.dataset.value);
+        const rightSideValue = parseFloat(rightStat.dataset.value);
 
-        if (rightSideValue > leftSideValue) {
+        if ( rightSideValue > leftSideValue) {
             leftStat.classList.remove('is-primary');
             leftStat.classList.add('is-warning');
             rightStat.classList.add('is-primary');
